refactor(hew-lsp): split server.rs into focused submodules
server.rs had grown to 5957 lines with 28 LSP handlers plus their
helper functions all in one flat file. The handlers were already thin
dispatchers; the bulk of the file was helper logic separated only by
// ── section comments.

Extract the helpers into five coherent submodules:

  server/analysis.rs   (666 lines)  – in-memory module resolution,
                                       document analysis lifecycle,
                                       diagnostic building from parse
                                       and type-checker output
  server/convert.rs    (163 lines)  – LSP type conversions: completion
                                       items, document symbols, semantic
                                       token encoding
  server/navigation.rs (828 lines)  – goto-definition, rename, find-
                                       references, import-path resolution,
                                       cross-file navigation, document links
  server/hierarchy.rs  (508 lines)  – type-hierarchy and call-hierarchy
                                       query helpers
  server/workspace.rs  (186 lines)  – code-lens construction and
                                       workspace symbol collection

server/mod.rs retains the shared types (DocumentState, DiagnosticMap,
DiagnosticSource), the coordinate helpers (position_to_offset,
span_to_range, offset_range_to_lsp), the semantic-token constants,
the LanguageServer trait impl, and the full test suite.

Each extracted function is pub(super); mod.rs uses explicit named
imports from each submodule (no wildcard re-exports) so clippy is
satisfied. Test-only helpers are gated with #[cfg(test)] imports.

No behavior changes; all 112 tests pass.
