+++ hew-lsp/src/server/mod.rs
@@ -1,39 +1,73 @@
 //! LSP server implementation for the Hew language.
 
-use std::collections::{HashMap, HashSet};
+mod analysis;
+mod convert;
+mod hierarchy;
+mod navigation;
+mod workspace;
+
+// Items used by the LanguageServer impl handlers.
+use self::analysis::{close_document_and_dependents, refresh_document_and_dependents};
+use self::convert::{analysis_tokens_to_lsp, symbol_info_to_doc_symbol, to_lsp_completion};
+use self::hierarchy::{
+    collect_subtypes, collect_supertypes, find_callable_at, find_incoming_calls,
+    find_outgoing_calls, find_type_hierarchy_item,
+};
+use self::navigation::{
+    build_document_links, build_prepare_rename_response, build_reference_locations,
+    build_workspace_edit, collect_import_items, find_cross_file_definition, find_definition_in_ast,
+};
+use self::workspace::{build_code_lenses, collect_workspace_symbols};
+
+// Items additionally needed by the test module (only compiled in test builds).
+#[cfg(test)]
+use self::analysis::{build_diagnostics, diagnostic_data, populate_user_module_imports};
+#[cfg(test)]
+use self::convert::analysis_symbol_kind_to_lsp;
+#[cfg(test)]
+use self::workspace::has_test_attribute;
+
+use std::collections::HashMap;
 
 use dashmap::DashMap;
-use hew_analysis::calls::{
-    collect_calls_in_block, collect_calls_in_item, collect_calls_in_named_body,
-};
+#[cfg(test)]
 use hew_analysis::references::count_all_references;
-use hew_analysis::util::{compute_line_offsets, non_empty, offset_to_line_col, word_at_offset};
-use hew_parser::ast::{Attribute, ImportDecl, ImportSpec, Item, Span, TypeDeclKind};
+#[cfg(test)]
+use hew_analysis::util::compute_line_offsets;
+use hew_analysis::util::{non_empty, offset_to_line_col, word_at_offset};
+use hew_parser::ast::Span;
+#[cfg(test)]
+use hew_parser::ast::{Attribute, Item};
 use hew_parser::ParseResult;
-use hew_types::error::{Severity, TypeErrorKind};
-use hew_types::module_registry::build_module_search_paths;
-use hew_types::{Checker, TypeCheckOutput};
+#[cfg(test)]
+use hew_types::error::TypeErrorKind;
+#[cfg(test)]
+use hew_types::Checker;
+use hew_types::TypeCheckOutput;
 use tower_lsp::jsonrpc::Result;
 use tower_lsp::lsp_types::{
     CallHierarchyIncomingCall, CallHierarchyIncomingCallsParams, CallHierarchyItem,
     CallHierarchyOutgoingCall, CallHierarchyOutgoingCallsParams, CallHierarchyPrepareParams,
-    CallHierarchyServerCapability, CodeLens, CodeLensOptions, CodeLensParams, Command,
-    SymbolInformation, WorkspaceSymbolParams,
+    CallHierarchyServerCapability, CodeLens, CodeLensOptions, CodeLensParams, SymbolInformation,
+    WorkspaceSymbolParams,
 };
 use tower_lsp::lsp_types::{
     CodeAction, CodeActionKind, CodeActionOrCommand, CodeActionParams, CodeActionResponse,
-    CompletionItem, CompletionItemKind, CompletionOptions, CompletionParams, CompletionResponse,
-    Diagnostic, DiagnosticRelatedInformation, DiagnosticSeverity, DidChangeTextDocumentParams,
-    DidCloseTextDocumentParams, DidOpenTextDocumentParams, DocumentSymbol, DocumentSymbolParams,
-    DocumentSymbolResponse, FoldingRange, FoldingRangeKind, FoldingRangeParams,
-    GotoDefinitionParams, GotoDefinitionResponse, Hover, HoverContents, HoverParams,
-    HoverProviderCapability, InitializeParams, InitializeResult, InitializedParams,
-    InsertTextFormat, Location, MarkupContent, MarkupKind, MessageType, OneOf, Position,
-    PrepareRenameResponse, Range, ReferenceParams, RenameParams, SemanticToken,
-    SemanticTokenModifier, SemanticTokenType, SemanticTokens, SemanticTokensFullOptions,
-    SemanticTokensLegend, SemanticTokensOptions, SemanticTokensParams, SemanticTokensResult,
-    SemanticTokensServerCapabilities, ServerCapabilities, SymbolKind, TextDocumentSyncCapability,
-    TextDocumentSyncKind, TextEdit, Url, WorkDoneProgressOptions, WorkspaceEdit,
+    CompletionOptions, CompletionParams, CompletionResponse, Diagnostic,
+    DidChangeTextDocumentParams, DidCloseTextDocumentParams, DidOpenTextDocumentParams,
+    DocumentSymbol, DocumentSymbolParams, DocumentSymbolResponse, FoldingRange, FoldingRangeKind,
+    FoldingRangeParams, GotoDefinitionParams, GotoDefinitionResponse, Hover, HoverContents,
+    HoverParams, HoverProviderCapability, InitializeParams, InitializeResult, InitializedParams,
+    Location, MarkupContent, MarkupKind, MessageType, OneOf, Position, PrepareRenameResponse,
+    Range, ReferenceParams, RenameParams, SemanticTokenModifier, SemanticTokenType, SemanticTokens,
+    SemanticTokensFullOptions, SemanticTokensLegend, SemanticTokensOptions, SemanticTokensParams,
+    SemanticTokensResult, SemanticTokensServerCapabilities, ServerCapabilities,
+    TextDocumentSyncCapability, TextDocumentSyncKind, TextEdit, Url, WorkDoneProgressOptions,
+    WorkspaceEdit,
+};
+#[cfg(test)]
+use tower_lsp::lsp_types::{
+    CompletionItemKind, DiagnosticSeverity, InsertTextFormat, SemanticToken, SymbolKind,
 };
 use tower_lsp::lsp_types::{DocumentLink, DocumentLinkOptions, DocumentLinkParams};
 use tower_lsp::lsp_types::{
@@ -94,519 +128,26 @@ struct DiagnosticSource {
     line_offsets: Vec<usize>,
 }
 
-// ── In-memory module resolution ──────────────────────────────────────
+// ── Helpers ──────────────────────────────────────────────────────────
 
-/// Return the source text for a file, preferring open editor buffers over disk.
-///
-/// Checks the LSP document store first so that unsaved edits to an imported
-/// module are immediately visible to type-checking in the importing file.
-/// Falls back to `std::fs::read_to_string` for files that are not open.
-fn source_for_path(
-    path: &std::path::Path,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<String> {
-    // Prefer in-memory content if the file is currently open in the editor.
-    if let Ok(url) = Url::from_file_path(path) {
-        if let Some(doc) = documents.get(&url) {
-            return Some(doc.source.clone());
-        }
+/// Convert an LSP `Position` (UTF-16 character offset) to a byte offset in source,
+/// using pre-computed line offsets.
+fn position_to_offset(source: &str, lo: &[usize], position: Position) -> usize {
+    hew_analysis::util::position_to_offset(source, lo, position.line, position.character)
+}
+
+/// Convert a parser `Span` (byte offsets) to an LSP `Range`.
+fn span_to_range(source: &str, lo: &[usize], span: &Span) -> Range {
+    offset_range_to_lsp(source, lo, span.start, span.end)
+}
+
+/// Convert byte offset range to LSP `Range`, using pre-computed line offsets.
+fn offset_range_to_lsp(source: &str, lo: &[usize], start: usize, end: usize) -> Range {
+    let (sl, sc, el, ec) = hew_analysis::util::span_to_line_col_range(source, lo, start, end);
+    Range {
+        start: Position::new(sl, sc),
+        end: Position::new(el, ec),
     }
-    // Fall back to on-disk content.
-    std::fs::read_to_string(path).ok()
-}
-
-/// Recursively populate `ImportDecl::resolved_items` for user and file imports.
-///
-/// After parsing a document the `resolved_items` field on every `ImportDecl`
-/// is `None`.  The CLI fills these in by reading files from disk; the LSP
-/// historically skipped this step, so imported modules were only resolved via
-/// the stdlib `ModuleRegistry` and not from the project tree.
-///
-/// This function walks `items`, finds unresolved `Import` nodes, locates the
-/// corresponding `.hew` file relative to `source_dir`, reads it — **preferring
-/// any open editor buffer over the on-disk version** — and populates
-/// `resolved_items` so the type checker sees the current in-memory content.
-///
-/// Depth is capped at [`MAX_IMPORT_DEPTH`] to prevent cycles.
-fn populate_user_module_imports(
-    source_uri: &Url,
-    items: &mut [hew_parser::ast::Spanned<hew_parser::ast::Item>],
-    documents: &DashMap<Url, DocumentState>,
-) {
-    let Ok(source_path) = source_uri.to_file_path() else {
-        return; // Non-file URI — nothing to resolve.
-    };
-    let Some(source_dir) = source_path.parent() else {
-        return;
-    };
-    populate_user_module_imports_impl(source_dir, items, documents, 0);
-}
-
-/// Maximum import nesting depth to prevent unbounded recursion on cycles.
-const MAX_IMPORT_DEPTH: usize = 16;
-
-fn populate_user_module_imports_impl(
-    source_dir: &std::path::Path,
-    items: &mut [hew_parser::ast::Spanned<hew_parser::ast::Item>],
-    documents: &DashMap<Url, DocumentState>,
-    depth: usize,
-) {
-    if depth >= MAX_IMPORT_DEPTH {
-        return;
-    }
-
-    for (item, _span) in items.iter_mut() {
-        let decl = match item {
-            // Only process imports that haven't been resolved yet.
-            Item::Import(d)
-                if d.resolved_items.is_none() && (d.file_path.is_some() || !d.path.is_empty()) =>
-            {
-                d
-            }
-            _ => continue,
-        };
-
-        let candidates = import_candidate_paths_from_dir(source_dir, decl);
-
-        for candidate in &candidates {
-            // `source_for_path` checks the in-memory document store first so
-            // unsaved edits are preferred over the on-disk version.
-            if let Some(source) = source_for_path(candidate, documents) {
-                let parsed = hew_parser::parse(&source);
-                let has_errors = parsed
-                    .errors
-                    .iter()
-                    .any(|e| e.severity == hew_parser::Severity::Error);
-                if !has_errors {
-                    let mut module_items = parsed.program.items;
-                    // Recursively resolve any imports inside the loaded module.
-                    let module_dir = candidate.parent().unwrap_or(source_dir);
-                    populate_user_module_imports_impl(
-                        module_dir,
-                        &mut module_items,
-                        documents,
-                        depth + 1,
-                    );
-                    let item_count = module_items.len();
-                    decl.resolved_source_paths = vec![candidate.clone()];
-                    decl.resolved_item_source_paths = vec![candidate.clone(); item_count];
-                    decl.resolved_items = Some(module_items);
-                }
-                // Stop after the first candidate that yielded source text,
-                // regardless of whether it parsed cleanly — otherwise we'd
-                // silently fall through to a stale on-disk version.
-                break;
-            }
-        }
-    }
-}
-
-fn import_candidate_paths_from_dir(
-    source_dir: &std::path::Path,
-    import: &ImportDecl,
-) -> Vec<std::path::PathBuf> {
-    if let Some(file_path) = &import.file_path {
-        return vec![source_dir.join(file_path)];
-    }
-
-    let Some(last) = import.path.last() else {
-        return vec![];
-    };
-
-    // Build the two canonical candidate paths the CLI also tries:
-    //   1. package-directory form:  source_dir/<a>/<b>/<b>.hew
-    //   2. flat form:               source_dir/<a>/<b>.hew
-    let rel_path: std::path::PathBuf = import
-        .path
-        .iter()
-        .collect::<std::path::PathBuf>()
-        .with_extension("hew");
-    let dir_path: std::path::PathBuf = import
-        .path
-        .iter()
-        .collect::<std::path::PathBuf>()
-        .join(format!("{last}.hew"));
-
-    vec![source_dir.join(&dir_path), source_dir.join(&rel_path)]
-}
-
-fn import_candidate_paths(uri: &Url, import: &ImportDecl) -> Vec<std::path::PathBuf> {
-    let Ok(source_path) = uri.to_file_path() else {
-        return vec![];
-    };
-    let Some(source_dir) = source_path.parent() else {
-        return vec![];
-    };
-
-    import_candidate_paths_from_dir(source_dir, import)
-}
-
-fn module_id_from_file(
-    source_dir: &std::path::Path,
-    canonical_path: &std::path::Path,
-) -> hew_parser::module::ModuleId {
-    use hew_parser::module::ModuleId;
-
-    let without_ext = canonical_path.with_extension("");
-    let rel = without_ext.strip_prefix(source_dir).unwrap_or(&without_ext);
-    let mut segments: Vec<String> = rel
-        .iter()
-        .filter_map(|segment| segment.to_str())
-        .map(std::string::ToString::to_string)
-        .collect();
-
-    if segments.is_empty() {
-        segments.push(
-            canonical_path
-                .file_stem()
-                .and_then(|segment| segment.to_str())
-                .unwrap_or("unknown")
-                .to_string(),
-        );
-    }
-
-    ModuleId::new(segments)
-}
-
-fn resolved_import_source_path(
-    current_source: &std::path::Path,
-    source_dir: &std::path::Path,
-    decl: &ImportDecl,
-) -> Option<std::path::PathBuf> {
-    decl.resolved_source_paths.first().cloned().or_else(|| {
-        decl.file_path.as_ref().map(|file_path| {
-            current_source
-                .parent()
-                .unwrap_or(source_dir)
-                .join(file_path)
-        })
-    })
-}
-
-fn build_document_module_graph(
-    source_uri: &Url,
-    program: &hew_parser::ast::Program,
-) -> Option<hew_parser::module::ModuleGraph> {
-    use hew_parser::module::{Module, ModuleGraph};
-
-    let input_path = source_uri.to_file_path().ok()?;
-    let input_path = std::fs::canonicalize(&input_path).unwrap_or(input_path);
-    let source_dir = input_path.parent().unwrap_or(std::path::Path::new("."));
-    let root_id = module_id_from_file(source_dir, &input_path);
-    let mut graph = ModuleGraph::new(root_id.clone());
-    let mut seen_ids = HashSet::from([root_id.clone()]);
-
-    let root_imports = extract_module_info(
-        &program.items,
-        &input_path,
-        source_dir,
-        &input_path,
-        &root_id,
-        &mut graph,
-        &mut seen_ids,
-    );
-
-    graph.add_module(Module {
-        id: root_id,
-        items: program.items.clone(),
-        imports: root_imports,
-        source_paths: vec![input_path],
-        doc: program.module_doc.clone(),
-    });
-
-    graph.compute_topo_order().ok()?;
-    Some(graph)
-}
-
-fn extract_module_info(
-    items: &[hew_parser::ast::Spanned<Item>],
-    current_source: &std::path::Path,
-    source_dir: &std::path::Path,
-    root_source: &std::path::Path,
-    root_id: &hew_parser::module::ModuleId,
-    graph: &mut hew_parser::module::ModuleGraph,
-    seen_ids: &mut HashSet<hew_parser::module::ModuleId>,
-) -> Vec<hew_parser::module::ModuleImport> {
-    use hew_parser::module::{Module, ModuleId, ModuleImport};
-
-    let mut imports = Vec::new();
-
-    for (item, span) in items {
-        let Item::Import(decl) = item else { continue };
-
-        let first_source_path = resolved_import_source_path(current_source, source_dir, decl);
-        let module_id = if !decl.path.is_empty() {
-            ModuleId::new(decl.path.clone())
-        } else if let Some(source_path) = first_source_path.as_ref() {
-            if source_path == root_source {
-                root_id.clone()
-            } else {
-                module_id_from_file(source_dir, source_path)
-            }
-        } else {
-            continue;
-        };
-
-        imports.push(ModuleImport {
-            target: module_id.clone(),
-            spec: decl.spec.clone(),
-            span: span.clone(),
-        });
-
-        if seen_ids.insert(module_id.clone()) {
-            if let Some(resolved_items) = &decl.resolved_items {
-                let child_source = first_source_path.as_deref().unwrap_or(current_source);
-                let child_imports = extract_module_info(
-                    resolved_items,
-                    child_source,
-                    source_dir,
-                    root_source,
-                    root_id,
-                    graph,
-                    seen_ids,
-                );
-                let source_paths = if decl.resolved_source_paths.is_empty() {
-                    first_source_path.iter().cloned().collect()
-                } else {
-                    decl.resolved_source_paths.clone()
-                };
-                graph.add_module(Module {
-                    id: module_id,
-                    items: resolved_items.clone(),
-                    imports: child_imports,
-                    source_paths,
-                    doc: None,
-                });
-            }
-        }
-    }
-
-    imports
-}
-
-fn build_module_source_map(
-    program: &hew_parser::ast::Program,
-    documents: &DashMap<Url, DocumentState>,
-) -> HashMap<String, DiagnosticSource> {
-    let Some(ref module_graph) = program.module_graph else {
-        return HashMap::new();
-    };
-
-    let mut module_sources = HashMap::new();
-    for module_id in &module_graph.topo_order {
-        if *module_id == module_graph.root {
-            continue;
-        }
-        let Some(module) = module_graph.modules.get(module_id) else {
-            continue;
-        };
-        let Some(source_path) = module.source_paths.first() else {
-            continue;
-        };
-        let Ok(uri) = Url::from_file_path(source_path) else {
-            continue;
-        };
-        let Some(source) = source_for_path(source_path, documents) else {
-            continue;
-        };
-        module_sources.insert(
-            module_id.path.join("."),
-            DiagnosticSource {
-                uri,
-                line_offsets: compute_line_offsets(&source),
-                source,
-            },
-        );
-    }
-
-    module_sources
-}
-
-fn merge_diagnostics(into: &mut DiagnosticMap, from: &DiagnosticMap) {
-    for (uri, diagnostics) in from {
-        into.entry(uri.clone())
-            .or_default()
-            .extend(diagnostics.iter().cloned());
-    }
-}
-
-fn sort_and_dedup_diagnostics(diagnostics: &mut Vec<Diagnostic>) {
-    diagnostics.sort_by(|left, right| {
-        left.range
-            .start
-            .line
-            .cmp(&right.range.start.line)
-            .then(left.range.start.character.cmp(&right.range.start.character))
-            .then(left.range.end.line.cmp(&right.range.end.line))
-            .then(left.range.end.character.cmp(&right.range.end.character))
-            .then(
-                left.source
-                    .as_deref()
-                    .unwrap_or("")
-                    .cmp(right.source.as_deref().unwrap_or("")),
-            )
-            .then(left.message.cmp(&right.message))
-    });
-
-    let mut seen = HashSet::new();
-    diagnostics.retain(|diagnostic| seen.insert(format!("{diagnostic:?}")));
-}
-
-fn collect_published_diagnostics(
-    documents: &DashMap<Url, DocumentState>,
-    publish_uris: HashSet<Url>,
-) -> Vec<(Url, Vec<Diagnostic>)> {
-    let mut diagnostics_by_uri = DiagnosticMap::new();
-    for entry in documents {
-        merge_diagnostics(&mut diagnostics_by_uri, &entry.value().diagnostics_by_uri);
-    }
-
-    for uri in publish_uris {
-        diagnostics_by_uri.entry(uri).or_default();
-    }
-
-    let mut published: Vec<_> = diagnostics_by_uri.into_iter().collect();
-    for (_, diagnostics) in &mut published {
-        sort_and_dedup_diagnostics(diagnostics);
-    }
-    published.sort_by(|(left_uri, _), (right_uri, _)| left_uri.as_str().cmp(right_uri.as_str()));
-    published
-}
-
-fn analyze_document(
-    uri: &Url,
-    source: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> DocumentState {
-    let parse_result = hew_parser::parse(source);
-    let line_offsets = compute_line_offsets(source);
-
-    let has_parse_errors = parse_result
-        .errors
-        .iter()
-        .any(|e| e.severity == hew_parser::Severity::Error);
-
-    let (type_output, module_sources) = if has_parse_errors {
-        (None, HashMap::new())
-    } else {
-        let mut checker = Checker::new(hew_types::module_registry::ModuleRegistry::new(
-            build_module_search_paths(),
-        ));
-        // Clone the program so we can inject resolved_items for user-module
-        // imports without mutating the parse_result stored in DocumentState
-        // (other LSP features use the raw AST and do not need resolved_items).
-        let mut program = parse_result.program.clone();
-        populate_user_module_imports(uri, &mut program.items, documents);
-        program.module_graph = build_document_module_graph(uri, &program);
-        let module_sources = build_module_source_map(&program, documents);
-        (Some(checker.check_program(&program)), module_sources)
-    };
-
-    let diagnostics_by_uri = build_diagnostics_by_uri(
-        uri,
-        source,
-        &line_offsets,
-        &parse_result,
-        type_output.as_ref(),
-        &module_sources,
-    );
-
-    DocumentState {
-        source: source.to_string(),
-        line_offsets,
-        parse_result,
-        type_output,
-        diagnostics_by_uri,
-    }
-}
-
-fn document_imports_target(
-    importer_uri: &Url,
-    parse_result: &ParseResult,
-    target_uri: &Url,
-) -> bool {
-    parse_result.program.items.iter().any(|(item, _)| {
-        let Item::Import(import) = item else {
-            return false;
-        };
-
-        import_candidate_paths(importer_uri, import)
-            .into_iter()
-            .filter_map(|path| Url::from_file_path(path).ok())
-            .any(|candidate_uri| candidate_uri == *target_uri)
-    })
-}
-
-fn refresh_open_importers(
-    target_uri: &Url,
-    documents: &DashMap<Url, DocumentState>,
-    publish_uris: &mut HashSet<Url>,
-) {
-    let dependents: Vec<_> = documents
-        .iter()
-        .filter_map(|entry| {
-            let importer_uri = entry.key().clone();
-            if &importer_uri == target_uri {
-                return None;
-            }
-
-            let importer = entry.value();
-            if document_imports_target(&importer_uri, &importer.parse_result, target_uri) {
-                Some((
-                    importer_uri,
-                    importer.source.clone(),
-                    importer
-                        .diagnostics_by_uri
-                        .keys()
-                        .cloned()
-                        .collect::<Vec<_>>(),
-                ))
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    for (importer_uri, importer_source, previous_diagnostic_uris) in dependents {
-        publish_uris.insert(importer_uri.clone());
-        publish_uris.extend(previous_diagnostic_uris);
-        let document = analyze_document(&importer_uri, &importer_source, documents);
-        documents.insert(importer_uri.clone(), document);
-    }
-}
-
-fn refresh_document_and_dependents(
-    uri: &Url,
-    source: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Vec<(Url, Vec<Diagnostic>)> {
-    let mut publish_uris = HashSet::from([uri.clone()]);
-    if let Some(previous) = documents.get(uri) {
-        publish_uris.extend(previous.diagnostics_by_uri.keys().cloned());
-    }
-
-    let document = analyze_document(uri, source, documents);
-    documents.insert(uri.clone(), document);
-
-    refresh_open_importers(uri, documents, &mut publish_uris);
-
-    collect_published_diagnostics(documents, publish_uris)
-}
-
-fn close_document_and_dependents(
-    uri: &Url,
-    documents: &DashMap<Url, DocumentState>,
-) -> Vec<(Url, Vec<Diagnostic>)> {
-    let Some((_removed_uri, removed_document)) = documents.remove(uri) else {
-        return vec![];
-    };
-
-    let mut publish_uris = HashSet::from([uri.clone()]);
-    publish_uris.extend(removed_document.diagnostics_by_uri.keys().cloned());
-
-    refresh_open_importers(uri, documents, &mut publish_uris);
-
-    collect_published_diagnostics(documents, publish_uris)
 }
 
 // ── Server ───────────────────────────────────────────────────────────
@@ -1304,1812 +845,6 @@ impl LanguageServer for HewLanguageServer {
             .collect();
         Ok(non_empty(lsp_ranges))
     }
-}
-
-// ── Diagnostics ──────────────────────────────────────────────────────
-
-fn insert_diagnostic(diagnostics_by_uri: &mut DiagnosticMap, uri: Url, diagnostic: Diagnostic) {
-    diagnostics_by_uri.entry(uri).or_default().push(diagnostic);
-}
-
-#[cfg(test)]
-fn build_diagnostics(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    type_output: Option<&TypeCheckOutput>,
-) -> Vec<Diagnostic> {
-    build_diagnostics_by_uri(uri, source, lo, parse_result, type_output, &HashMap::new())
-        .remove(uri)
-        .unwrap_or_default()
-}
-
-fn build_diagnostics_by_uri(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    type_output: Option<&TypeCheckOutput>,
-    module_sources: &HashMap<String, DiagnosticSource>,
-) -> DiagnosticMap {
-    let mut diagnostics_by_uri = DiagnosticMap::new();
-
-    for err in &parse_result.errors {
-        let message = if let Some(hint) = &err.hint {
-            format!("{}\n\nhint: {hint}", err.message)
-        } else {
-            err.message.clone()
-        };
-        let severity = match err.severity {
-            hew_parser::Severity::Error => DiagnosticSeverity::ERROR,
-            hew_parser::Severity::Warning => DiagnosticSeverity::WARNING,
-        };
-        insert_diagnostic(
-            &mut diagnostics_by_uri,
-            uri.clone(),
-            Diagnostic {
-                range: span_to_range(source, lo, &err.span),
-                severity: Some(severity),
-                source: Some("hew-parser".to_string()),
-                message,
-                ..Default::default()
-            },
-        );
-    }
-
-    if let Some(tc) = type_output {
-        for diag in tc.errors.iter().chain(tc.warnings.iter()) {
-            let target = diag
-                .source_module
-                .as_ref()
-                .and_then(|module_name| module_sources.get(module_name));
-            let (target_uri, target_source, target_line_offsets) = if let Some(target) = target {
-                (
-                    target.uri.clone(),
-                    target.source.as_str(),
-                    target.line_offsets.as_slice(),
-                )
-            } else {
-                (uri.clone(), source, lo)
-            };
-
-            // Build the message, appending any suggestions.
-            let message = if diag.suggestions.is_empty() {
-                diag.message.clone()
-            } else {
-                format!(
-                    "{}\n\nDid you mean: {}",
-                    diag.message,
-                    diag.suggestions.join(", ")
-                )
-            };
-
-            // Surface notes as related information with their own spans.
-            let related_information = if diag.notes.is_empty() {
-                None
-            } else {
-                Some(
-                    diag.notes
-                        .iter()
-                        .map(|(note_span, note_msg)| DiagnosticRelatedInformation {
-                            location: Location {
-                                uri: target_uri.clone(),
-                                range: span_to_range(target_source, target_line_offsets, note_span),
-                            },
-                            message: note_msg.clone(),
-                        })
-                        .collect(),
-                )
-            };
-
-            insert_diagnostic(
-                &mut diagnostics_by_uri,
-                target_uri,
-                Diagnostic {
-                    range: span_to_range(target_source, target_line_offsets, &diag.span),
-                    severity: Some(severity_to_lsp(diag.severity)),
-                    source: Some("hew-types".to_string()),
-                    message,
-                    related_information,
-                    data: Some(diagnostic_data(&diag.kind, &diag.suggestions)),
-                    ..Default::default()
-                },
-            );
-        }
-    }
-
-    diagnostics_by_uri
-}
-
-/// Convert a type-checker `Severity` to the corresponding LSP `DiagnosticSeverity`.
-///
-/// This is the authoritative severity mapping: the `TypeError` struct carries the
-/// severity that was decided at emit time, so the LSP should honour it directly
-/// rather than re-deriving it from the `TypeErrorKind`.
-fn severity_to_lsp(severity: Severity) -> DiagnosticSeverity {
-    match severity {
-        Severity::Error => DiagnosticSeverity::ERROR,
-        Severity::Warning => DiagnosticSeverity::WARNING,
-    }
-}
-
-/// Encode a `TypeErrorKind` discriminant and suggestions as JSON for `Diagnostic.data`.
-fn diagnostic_data(kind: &TypeErrorKind, suggestions: &[String]) -> serde_json::Value {
-    serde_json::json!({
-        "kind": kind.as_kind_str(),
-        "suggestions": suggestions,
-    })
-}
-
-// ── Completion ───────────────────────────────────────────────────────
-
-/// Convert an `hew_analysis::CompletionItem` to an `lsp_types::CompletionItem`.
-fn to_lsp_completion(item: hew_analysis::CompletionItem) -> CompletionItem {
-    use hew_analysis::CompletionKind;
-    let kind = match item.kind {
-        CompletionKind::Function => CompletionItemKind::FUNCTION,
-        CompletionKind::Variable => CompletionItemKind::VARIABLE,
-        CompletionKind::Keyword => CompletionItemKind::KEYWORD,
-        CompletionKind::Snippet => CompletionItemKind::SNIPPET,
-        CompletionKind::Type => CompletionItemKind::STRUCT,
-        CompletionKind::Actor => CompletionItemKind::CLASS,
-        CompletionKind::Constant => CompletionItemKind::CONSTANT,
-        CompletionKind::Field => CompletionItemKind::FIELD,
-        CompletionKind::Method => CompletionItemKind::METHOD,
-        CompletionKind::Module => CompletionItemKind::MODULE,
-    };
-    CompletionItem {
-        label: item.label,
-        kind: Some(kind),
-        detail: item.detail,
-        insert_text: item.insert_text,
-        insert_text_format: if item.insert_text_is_snippet {
-            Some(InsertTextFormat::SNIPPET)
-        } else {
-            None
-        },
-        sort_text: item.sort_text,
-        ..Default::default()
-    }
-}
-
-// ── Document symbols ─────────────────────────────────────────────────
-
-/// Convert an `hew_analysis::SymbolKind` to an LSP `SymbolKind`.
-fn analysis_symbol_kind_to_lsp(kind: hew_analysis::SymbolKind) -> SymbolKind {
-    match kind {
-        hew_analysis::SymbolKind::Function => SymbolKind::FUNCTION,
-        hew_analysis::SymbolKind::Actor | hew_analysis::SymbolKind::Supervisor => SymbolKind::CLASS,
-        hew_analysis::SymbolKind::Machine | hew_analysis::SymbolKind::Enum => SymbolKind::ENUM,
-        hew_analysis::SymbolKind::Trait => SymbolKind::INTERFACE,
-        hew_analysis::SymbolKind::Type | hew_analysis::SymbolKind::Wire => SymbolKind::STRUCT,
-        hew_analysis::SymbolKind::Constant => SymbolKind::CONSTANT,
-        hew_analysis::SymbolKind::TypeAlias => SymbolKind::TYPE_PARAMETER,
-        hew_analysis::SymbolKind::Impl => SymbolKind::NAMESPACE,
-        hew_analysis::SymbolKind::Field => SymbolKind::FIELD,
-        hew_analysis::SymbolKind::Method => SymbolKind::METHOD,
-        hew_analysis::SymbolKind::State | hew_analysis::SymbolKind::Variant => {
-            SymbolKind::ENUM_MEMBER
-        }
-        hew_analysis::SymbolKind::Event => SymbolKind::EVENT,
-        hew_analysis::SymbolKind::Module => SymbolKind::MODULE,
-        hew_analysis::SymbolKind::Constructor => SymbolKind::CONSTRUCTOR,
-    }
-}
-
-/// Convert an `hew_analysis::SymbolInfo` to an LSP `DocumentSymbol`.
-#[expect(
-    deprecated,
-    reason = "DocumentSymbol::deprecated is required by the LSP type"
-)]
-fn symbol_info_to_doc_symbol(
-    source: &str,
-    lo: &[usize],
-    info: hew_analysis::SymbolInfo,
-) -> DocumentSymbol {
-    let range = offset_range_to_lsp(source, lo, info.span.start, info.span.end);
-    let selection_range = offset_range_to_lsp(
-        source,
-        lo,
-        info.selection_span.start,
-        info.selection_span.end,
-    );
-    let children = if info.children.is_empty() {
-        None
-    } else {
-        Some(
-            info.children
-                .into_iter()
-                .map(|c| symbol_info_to_doc_symbol(source, lo, c))
-                .collect(),
-        )
-    };
-    DocumentSymbol {
-        name: info.name,
-        detail: None,
-        kind: analysis_symbol_kind_to_lsp(info.kind),
-        tags: None,
-        deprecated: None,
-        range,
-        selection_range,
-        children,
-    }
-}
-
-// ── Semantic tokens ──────────────────────────────────────────────────
-
-/// Convert analysis semantic tokens (absolute byte offsets) to LSP
-/// delta-encoded tokens (line/col deltas, UTF-16 lengths).
-#[expect(
-    clippy::cast_possible_truncation,
-    reason = "token delta values in source files will not exceed u32"
-)]
-fn analysis_tokens_to_lsp(
-    source: &str,
-    lo: &[usize],
-    tokens: &[hew_analysis::SemanticToken],
-) -> Vec<SemanticToken> {
-    let mut result = Vec::new();
-    let mut prev_line: u32 = 0;
-    let mut prev_start: u32 = 0;
-
-    for tok in tokens {
-        let (line, col) = offset_to_line_col(source, lo, tok.start);
-        let line = line as u32;
-        let col = col as u32;
-        // UTF-16 length for the LSP protocol
-        let length: u32 = source[tok.start..tok.start + tok.length]
-            .chars()
-            .map(|c| c.len_utf16() as u32)
-            .sum();
-
-        let delta_line = line - prev_line;
-        let delta_start = if delta_line == 0 {
-            col - prev_start
-        } else {
-            col
-        };
-
-        // Map analysis modifier bitmasks to LSP modifier bitmasks.
-        let mut lsp_modifiers: u32 = 0;
-        if tok.modifiers & hew_analysis::token_modifiers::DECLARATION != 0 {
-            lsp_modifiers |= modifier_bit(&SemanticTokenModifier::DECLARATION);
-        }
-        if tok.modifiers & hew_analysis::token_modifiers::READONLY != 0 {
-            lsp_modifiers |= modifier_bit(&SemanticTokenModifier::READONLY);
-        }
-        if tok.modifiers & hew_analysis::token_modifiers::ASYNC != 0 {
-            lsp_modifiers |= modifier_bit(&SemanticTokenModifier::ASYNC);
-        }
-
-        result.push(SemanticToken {
-            delta_line,
-            delta_start,
-            length,
-            token_type: tok.token_type,
-            token_modifiers_bitset: lsp_modifiers,
-        });
-
-        prev_line = line;
-        prev_start = col;
-    }
-
-    result
-}
-
-// ── Helpers ──────────────────────────────────────────────────────────
-
-/// Convert an LSP `Position` (UTF-16 character offset) to a byte offset in source,
-/// using pre-computed line offsets.
-fn position_to_offset(source: &str, lo: &[usize], position: Position) -> usize {
-    hew_analysis::util::position_to_offset(source, lo, position.line, position.character)
-}
-
-/// Convert a parser `Span` (byte offsets) to an LSP `Range`.
-fn span_to_range(source: &str, lo: &[usize], span: &Span) -> Range {
-    offset_range_to_lsp(source, lo, span.start, span.end)
-}
-
-/// Convert byte offset range to LSP `Range`, using pre-computed line offsets.
-fn offset_range_to_lsp(source: &str, lo: &[usize], start: usize, end: usize) -> Range {
-    let (sl, sc, el, ec) = hew_analysis::util::span_to_line_col_range(source, lo, start, end);
-    Range {
-        start: Position::new(sl, sc),
-        end: Position::new(el, ec),
-    }
-}
-
-// ── Go-to-definition ─────────────────────────────────────────────────
-
-/// Search for a definition matching `word` in the AST, returning an LSP `Range`.
-fn find_definition_in_ast(
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    word: &str,
-) -> Option<Range> {
-    let span = hew_analysis::definition::find_definition(source, parse_result, word)?;
-    Some(offset_range_to_lsp(source, lo, span.start, span.end))
-}
-
-// ── Import path resolution ────────────────────────────────────────────
-
-/// Compute the absolute file path that an `ImportDecl` would resolve to,
-/// searching the workspace root first then the importing file's directory.
-///
-/// Does **not** check whether the file exists on disk; callers decide whether
-/// to check existence before performing I/O.
-fn compute_import_path(uri: &Url, import: &ImportDecl) -> Option<std::path::PathBuf> {
-    // String-literal import: `import "relative/path.hew";`
-    if let Some(fp) = &import.file_path {
-        let file_dir = uri
-            .to_file_path()
-            .ok()
-            .and_then(|p| p.parent().map(std::path::Path::to_path_buf))?;
-        return Some(file_dir.join(fp));
-    }
-
-    if import.path.is_empty() {
-        return None;
-    }
-
-    let relative = format!("{}.hew", import.path.join("/"));
-
-    // Workspace root: the directory that directly contains a `std/` folder.
-    let workspace_root = uri.to_file_path().ok().and_then(|p| {
-        let mut dir = p.parent().map(std::path::Path::to_path_buf);
-        while let Some(d) = dir {
-            if d.join("std").is_dir() {
-                return Some(d);
-            }
-            dir = d.parent().map(std::path::Path::to_path_buf);
-        }
-        None
-    });
-
-    // Prefer workspace root when the file already exists there.
-    if let Some(root) = workspace_root {
-        let candidate = root.join(&relative);
-        if candidate.exists() {
-            return Some(candidate);
-        }
-    }
-
-    // Fall back to the directory of the importing file.
-    let file_dir = uri
-        .to_file_path()
-        .ok()
-        .and_then(|p| p.parent().map(std::path::Path::to_path_buf))?;
-    Some(file_dir.join(&relative))
-}
-
-fn collect_import_items(parse_result: &ParseResult) -> Vec<(ImportDecl, Span)> {
-    parse_result
-        .program
-        .items
-        .iter()
-        .filter_map(|(item, span)| match item {
-            Item::Import(import) => Some((import.clone(), span.clone())),
-            _ => None,
-        })
-        .collect()
-}
-
-fn is_identifier_char(ch: char) -> bool {
-    ch.is_alphanumeric() || ch == '_'
-}
-
-fn find_identifier_span_in_range(
-    source: &str,
-    range: std::ops::Range<usize>,
-    ident: &str,
-) -> Option<hew_analysis::OffsetSpan> {
-    if ident.is_empty() {
-        return None;
-    }
-
-    let slice = source.get(range.clone())?;
-    let mut search_start = 0;
-    while let Some(relative_start) = slice[search_start..].find(ident) {
-        let start = range.start + search_start + relative_start;
-        let end = start + ident.len();
-        let prev_is_ident = source[..start]
-            .chars()
-            .next_back()
-            .is_some_and(is_identifier_char);
-        let next_is_ident = source[end..].chars().next().is_some_and(is_identifier_char);
-        if !prev_is_ident && !next_is_ident {
-            return Some(hew_analysis::OffsetSpan { start, end });
-        }
-        search_start += relative_start + ident.len();
-    }
-
-    None
-}
-
-fn find_named_import_spans(
-    source: &str,
-    item_span: &Span,
-    import_name: &hew_parser::ast::ImportName,
-) -> Option<(hew_analysis::OffsetSpan, hew_analysis::OffsetSpan)> {
-    let item_text = source.get(item_span.clone())?;
-    let open_brace = item_text.find('{')?;
-    let close_brace = item_text.rfind('}')?;
-    if open_brace >= close_brace {
-        return None;
-    }
-
-    let names_range = item_span.start + open_brace + 1..item_span.start + close_brace;
-    let import_name_span =
-        find_identifier_span_in_range(source, names_range.clone(), &import_name.name)?;
-    let visible_name_span = match &import_name.alias {
-        Some(alias) => find_identifier_span_in_range(source, names_range, alias)?,
-        None => import_name_span,
-    };
-
-    Some((import_name_span, visible_name_span))
-}
-
-/// Cross-file references/rename currently follow explicit named imports between
-/// open documents, reusing the existing import path resolution without a
-/// broader project index.
-#[derive(Debug, Clone)]
-struct NamedImportMatch {
-    importer_uri: Url,
-    imported_uri: Url,
-    imported_name: String,
-    visible_name: String,
-    import_name_span: hew_analysis::OffsetSpan,
-    visible_name_span: hew_analysis::OffsetSpan,
-}
-
-impl NamedImportMatch {
-    fn is_aliased(&self) -> bool {
-        self.visible_name != self.imported_name
-    }
-}
-
-fn find_named_import_match(
-    current_uri: &Url,
-    source: &str,
-    parse_result: &ParseResult,
-    word: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<NamedImportMatch> {
-    for (import, item_span) in collect_import_items(parse_result) {
-        let Some(ImportSpec::Names(names)) = &import.spec else {
-            continue;
-        };
-        let Some(path) = compute_import_path(current_uri, &import) else {
-            continue;
-        };
-        let Ok(imported_uri) = Url::from_file_path(&path) else {
-            continue;
-        };
-        let Some(target_doc) = documents.get(&imported_uri) else {
-            continue;
-        };
-
-        for import_name in names {
-            let visible_name = import_name
-                .alias
-                .as_deref()
-                .unwrap_or(import_name.name.as_str());
-            if visible_name != word {
-                continue;
-            }
-            if hew_analysis::definition::find_definition(
-                &target_doc.source,
-                &target_doc.parse_result,
-                &import_name.name,
-            )
-            .is_none()
-            {
-                continue;
-            }
-            let Some((import_name_span, visible_name_span)) =
-                find_named_import_spans(source, &item_span, import_name)
-            else {
-                continue;
-            };
-            return Some(NamedImportMatch {
-                importer_uri: current_uri.clone(),
-                imported_uri,
-                imported_name: import_name.name.clone(),
-                visible_name: visible_name.to_string(),
-                import_name_span,
-                visible_name_span,
-            });
-        }
-    }
-
-    None
-}
-
-fn span_contains_offset(span: hew_analysis::OffsetSpan, offset: usize) -> bool {
-    span.start <= offset && offset < span.end
-}
-
-fn find_resolved_named_import_match(
-    current_uri: &Url,
-    doc: &DocumentState,
-    offset: usize,
-    word: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<(NamedImportMatch, Vec<hew_analysis::OffsetSpan>)> {
-    let import_match =
-        find_named_import_match(current_uri, &doc.source, &doc.parse_result, word, documents)?;
-    let usage_spans = hew_analysis::references::find_import_binding_references(
-        &doc.parse_result,
-        &import_match.visible_name,
-    );
-
-    if span_contains_offset(import_match.visible_name_span, offset)
-        || usage_spans
-            .iter()
-            .any(|span| span_contains_offset(*span, offset))
-    {
-        Some((import_match, usage_spans))
-    } else {
-        None
-    }
-}
-
-fn find_open_named_importers(
-    target_uri: &Url,
-    target_name: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Vec<NamedImportMatch> {
-    let mut matches = Vec::new();
-
-    for entry in documents {
-        let importer_uri = entry.key().clone();
-        if importer_uri == *target_uri {
-            continue;
-        }
-        let doc = entry.value();
-
-        for (import, item_span) in collect_import_items(&doc.parse_result) {
-            let Some(ImportSpec::Names(names)) = &import.spec else {
-                continue;
-            };
-            let Some(path) = compute_import_path(&importer_uri, &import) else {
-                continue;
-            };
-            let Ok(resolved_uri) = Url::from_file_path(&path) else {
-                continue;
-            };
-            if resolved_uri != *target_uri {
-                continue;
-            }
-
-            for import_name in names {
-                if import_name.name != target_name {
-                    continue;
-                }
-                let Some((import_name_span, visible_name_span)) =
-                    find_named_import_spans(&doc.source, &item_span, import_name)
-                else {
-                    continue;
-                };
-                matches.push(NamedImportMatch {
-                    importer_uri: importer_uri.clone(),
-                    imported_uri: resolved_uri.clone(),
-                    imported_name: import_name.name.clone(),
-                    visible_name: import_name
-                        .alias
-                        .as_deref()
-                        .unwrap_or(import_name.name.as_str())
-                        .to_string(),
-                    import_name_span,
-                    visible_name_span,
-                });
-            }
-        }
-    }
-
-    matches
-}
-
-fn find_definition_name_span(
-    source: &str,
-    parse_result: &ParseResult,
-    name: &str,
-) -> Option<hew_analysis::OffsetSpan> {
-    let item_span = hew_analysis::definition::find_definition(source, parse_result, name)?;
-    find_identifier_span_in_range(source, item_span.start..item_span.end, name)
-}
-
-fn push_location_for_span(
-    locations: &mut Vec<Location>,
-    uri: &Url,
-    source: &str,
-    line_offsets: &[usize],
-    span: hew_analysis::OffsetSpan,
-) {
-    locations.push(Location {
-        uri: uri.clone(),
-        range: offset_range_to_lsp(source, line_offsets, span.start, span.end),
-    });
-}
-
-fn collect_local_reference_locations(
-    uri: &Url,
-    doc: &DocumentState,
-    offset: usize,
-    include_declaration: bool,
-) -> Vec<Location> {
-    let mut locations = Vec::new();
-
-    if let Some((_name, spans)) =
-        hew_analysis::references::find_all_references(&doc.source, &doc.parse_result, offset)
-    {
-        for span in spans {
-            push_location_for_span(&mut locations, uri, &doc.source, &doc.line_offsets, span);
-        }
-    }
-
-    if include_declaration {
-        if let Some((name, _)) = hew_analysis::util::simple_word_at_offset(&doc.source, offset) {
-            if let Some(def_span) =
-                hew_analysis::definition::find_definition(&doc.source, &doc.parse_result, &name)
-            {
-                push_location_for_span(
-                    &mut locations,
-                    uri,
-                    &doc.source,
-                    &doc.line_offsets,
-                    def_span,
-                );
-            }
-        }
-    }
-
-    locations
-}
-
-fn build_prepare_rename_response(
-    uri: &Url,
-    doc: &DocumentState,
-    offset: usize,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<PrepareRenameResponse> {
-    let span = if let Some((name, span)) =
-        hew_analysis::util::simple_word_at_offset(&doc.source, offset)
-    {
-        if find_resolved_named_import_match(uri, doc, offset, &name, documents).is_some() {
-            span
-        } else {
-            hew_analysis::rename::prepare_rename(&doc.source, &doc.parse_result, offset)?
-        }
-    } else {
-        hew_analysis::rename::prepare_rename(&doc.source, &doc.parse_result, offset)?
-    };
-    let range = offset_range_to_lsp(&doc.source, &doc.line_offsets, span.start, span.end);
-    Some(PrepareRenameResponse::Range(range))
-}
-
-fn sort_and_dedup_locations(locations: &mut Vec<Location>) {
-    locations.sort_by(|left, right| {
-        left.uri
-            .as_str()
-            .cmp(right.uri.as_str())
-            .then(left.range.start.line.cmp(&right.range.start.line))
-            .then(left.range.start.character.cmp(&right.range.start.character))
-            .then(left.range.end.line.cmp(&right.range.end.line))
-            .then(left.range.end.character.cmp(&right.range.end.character))
-    });
-    locations.dedup_by(|left, right| left.uri == right.uri && left.range == right.range);
-}
-
-fn collect_local_rename_edits(
-    doc: &DocumentState,
-    offset: usize,
-    new_name: &str,
-) -> Vec<hew_analysis::RenameEdit> {
-    hew_analysis::rename::rename(&doc.source, &doc.parse_result, offset, new_name)
-        .unwrap_or_default()
-}
-
-fn sort_and_dedup_rename_edits(edits: &mut Vec<hew_analysis::RenameEdit>) {
-    edits.sort_by(|left, right| {
-        left.span
-            .start
-            .cmp(&right.span.start)
-            .then(left.span.end.cmp(&right.span.end))
-            .then_with(|| left.new_text.cmp(&right.new_text))
-    });
-    edits.dedup_by(|left, right| left.span == right.span && left.new_text == right.new_text);
-}
-
-fn rename_edit_to_text_edit(doc: &DocumentState, edit: hew_analysis::RenameEdit) -> TextEdit {
-    TextEdit {
-        range: offset_range_to_lsp(
-            &doc.source,
-            &doc.line_offsets,
-            edit.span.start,
-            edit.span.end,
-        ),
-        new_text: edit.new_text,
-    }
-}
-
-fn workspace_edit_from_changes(
-    uri: &Url,
-    doc: &DocumentState,
-    documents: &DashMap<Url, DocumentState>,
-    mut changes: HashMap<Url, Vec<hew_analysis::RenameEdit>>,
-) -> Option<WorkspaceEdit> {
-    for edits in changes.values_mut() {
-        sort_and_dedup_rename_edits(edits);
-    }
-    changes.retain(|_, edits| !edits.is_empty());
-
-    if changes.is_empty() {
-        return None;
-    }
-
-    let mut lsp_changes = HashMap::new();
-    for (target_uri, edits) in changes {
-        let text_edits = if target_uri == *uri {
-            edits
-                .into_iter()
-                .map(|edit| rename_edit_to_text_edit(doc, edit))
-                .collect()
-        } else if let Some(target_doc) = documents.get(&target_uri) {
-            edits
-                .into_iter()
-                .map(|edit| rename_edit_to_text_edit(&target_doc, edit))
-                .collect()
-        } else {
-            continue;
-        };
-        lsp_changes.insert(target_uri, text_edits);
-    }
-
-    Some(WorkspaceEdit {
-        changes: Some(lsp_changes),
-        ..Default::default()
-    })
-}
-
-fn build_reference_locations(
-    uri: &Url,
-    doc: &DocumentState,
-    offset: usize,
-    include_declaration: bool,
-    documents: &DashMap<Url, DocumentState>,
-) -> Vec<Location> {
-    let Some((name, _)) = hew_analysis::util::simple_word_at_offset(&doc.source, offset) else {
-        return Vec::new();
-    };
-
-    if let Some((import_match, usage_spans)) =
-        find_resolved_named_import_match(uri, doc, offset, &name, documents)
-    {
-        let mut locations = Vec::new();
-        push_location_for_span(
-            &mut locations,
-            &import_match.importer_uri,
-            &doc.source,
-            &doc.line_offsets,
-            import_match.visible_name_span,
-        );
-        for span in usage_spans {
-            push_location_for_span(
-                &mut locations,
-                &import_match.importer_uri,
-                &doc.source,
-                &doc.line_offsets,
-                span,
-            );
-        }
-
-        if let Some(target_doc) = documents.get(&import_match.imported_uri) {
-            if let Some(def_span) = find_definition_name_span(
-                &target_doc.source,
-                &target_doc.parse_result,
-                &import_match.imported_name,
-            ) {
-                locations.extend(collect_local_reference_locations(
-                    &import_match.imported_uri,
-                    &target_doc,
-                    def_span.start,
-                    include_declaration,
-                ));
-            }
-        }
-
-        for importer in find_open_named_importers(
-            &import_match.imported_uri,
-            &import_match.imported_name,
-            documents,
-        ) {
-            if importer.importer_uri == import_match.importer_uri {
-                continue;
-            }
-            if let Some(importer_doc) = documents.get(&importer.importer_uri) {
-                push_location_for_span(
-                    &mut locations,
-                    &importer.importer_uri,
-                    &importer_doc.source,
-                    &importer_doc.line_offsets,
-                    importer.import_name_span,
-                );
-                for span in hew_analysis::references::find_import_binding_references(
-                    &importer_doc.parse_result,
-                    &importer.visible_name,
-                ) {
-                    push_location_for_span(
-                        &mut locations,
-                        &importer.importer_uri,
-                        &importer_doc.source,
-                        &importer_doc.line_offsets,
-                        span,
-                    );
-                }
-            }
-        }
-
-        sort_and_dedup_locations(&mut locations);
-        return locations;
-    }
-
-    let mut locations = collect_local_reference_locations(uri, doc, offset, include_declaration);
-
-    if hew_analysis::references::is_top_level_name(&doc.parse_result, &name) {
-        for importer in find_open_named_importers(uri, &name, documents) {
-            if let Some(importer_doc) = documents.get(&importer.importer_uri) {
-                push_location_for_span(
-                    &mut locations,
-                    &importer.importer_uri,
-                    &importer_doc.source,
-                    &importer_doc.line_offsets,
-                    importer.import_name_span,
-                );
-                for span in hew_analysis::references::find_import_binding_references(
-                    &importer_doc.parse_result,
-                    &importer.visible_name,
-                ) {
-                    push_location_for_span(
-                        &mut locations,
-                        &importer.importer_uri,
-                        &importer_doc.source,
-                        &importer_doc.line_offsets,
-                        span,
-                    );
-                }
-            }
-        }
-    }
-
-    sort_and_dedup_locations(&mut locations);
-    locations
-}
-
-#[expect(
-    clippy::too_many_lines,
-    reason = "cross-file rename handles local, imported-definition, and importer fanout paths"
-)]
-fn build_workspace_edit(
-    uri: &Url,
-    doc: &DocumentState,
-    offset: usize,
-    new_name: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<WorkspaceEdit> {
-    let (name, _) = hew_analysis::util::simple_word_at_offset(&doc.source, offset)?;
-
-    if let Some((import_match, usage_spans)) =
-        find_resolved_named_import_match(uri, doc, offset, &name, documents)
-    {
-        let mut changes: HashMap<Url, Vec<hew_analysis::RenameEdit>> = HashMap::new();
-        let mut importer_edits: Vec<_> = usage_spans
-            .into_iter()
-            .map(|span| hew_analysis::RenameEdit {
-                span,
-                new_text: new_name.to_string(),
-            })
-            .collect();
-        importer_edits.push(hew_analysis::RenameEdit {
-            span: import_match.visible_name_span,
-            new_text: new_name.to_string(),
-        });
-        changes.insert(uri.clone(), importer_edits);
-
-        if !import_match.is_aliased() {
-            if let Some(target_doc) = documents.get(&import_match.imported_uri) {
-                if let Some(def_span) = find_definition_name_span(
-                    &target_doc.source,
-                    &target_doc.parse_result,
-                    &import_match.imported_name,
-                ) {
-                    let target_edits =
-                        collect_local_rename_edits(&target_doc, def_span.start, new_name);
-                    if !target_edits.is_empty() {
-                        changes
-                            .entry(import_match.imported_uri.clone())
-                            .or_default()
-                            .extend(target_edits);
-                    }
-                }
-            }
-
-            for importer in find_open_named_importers(
-                &import_match.imported_uri,
-                &import_match.imported_name,
-                documents,
-            ) {
-                if importer.importer_uri == *uri {
-                    continue;
-                }
-                changes
-                    .entry(importer.importer_uri.clone())
-                    .or_default()
-                    .push(hew_analysis::RenameEdit {
-                        span: importer.import_name_span,
-                        new_text: new_name.to_string(),
-                    });
-
-                if importer.is_aliased() {
-                    continue;
-                }
-
-                if let Some(importer_doc) = documents.get(&importer.importer_uri) {
-                    let importer_edits: Vec<_> =
-                        hew_analysis::references::find_import_binding_references(
-                            &importer_doc.parse_result,
-                            &importer.visible_name,
-                        )
-                        .into_iter()
-                        .map(|span| hew_analysis::RenameEdit {
-                            span,
-                            new_text: new_name.to_string(),
-                        })
-                        .collect();
-                    if !importer_edits.is_empty() {
-                        changes
-                            .entry(importer.importer_uri.clone())
-                            .or_default()
-                            .extend(importer_edits);
-                    }
-                }
-            }
-        }
-
-        return workspace_edit_from_changes(uri, doc, documents, changes);
-    }
-
-    let mut changes: HashMap<Url, Vec<hew_analysis::RenameEdit>> = HashMap::new();
-    let local_edits = collect_local_rename_edits(doc, offset, new_name);
-    if !local_edits.is_empty() {
-        changes.insert(uri.clone(), local_edits);
-    }
-
-    if hew_analysis::references::is_top_level_name(&doc.parse_result, &name) {
-        for importer in find_open_named_importers(uri, &name, documents) {
-            changes
-                .entry(importer.importer_uri.clone())
-                .or_default()
-                .push(hew_analysis::RenameEdit {
-                    span: importer.import_name_span,
-                    new_text: new_name.to_string(),
-                });
-
-            if importer.is_aliased() {
-                continue;
-            }
-
-            if let Some(importer_doc) = documents.get(&importer.importer_uri) {
-                let importer_edits: Vec<_> =
-                    hew_analysis::references::find_import_binding_references(
-                        &importer_doc.parse_result,
-                        &importer.visible_name,
-                    )
-                    .into_iter()
-                    .map(|span| hew_analysis::RenameEdit {
-                        span,
-                        new_text: new_name.to_string(),
-                    })
-                    .collect();
-                if !importer_edits.is_empty() {
-                    changes
-                        .entry(importer.importer_uri.clone())
-                        .or_default()
-                        .extend(importer_edits);
-                }
-            }
-        }
-    }
-
-    workspace_edit_from_changes(uri, doc, documents, changes)
-}
-
-// ── Cross-file go-to-definition ───────────────────────────────────────
-
-/// Search for the definition of `word` in the files imported by the current
-/// file's `parse_result`.
-///
-/// Resolution order for each matching import:
-/// 1. Open documents already held by the server (no disk I/O).
-/// 2. The file on disk (parsed on demand).
-///
-/// Returns `(target_uri, range)` for the first match found, or `None`.
-fn find_cross_file_definition(
-    current_uri: &Url,
-    imports: &[hew_parser::ast::ImportDecl],
-    word: &str,
-    documents: &DashMap<Url, DocumentState>,
-) -> Option<(Url, Range)> {
-    for import in imports {
-        let Some(path) = compute_import_path(current_uri, import) else {
-            continue;
-        };
-        let Ok(target_uri) = Url::from_file_path(&path) else {
-            continue;
-        };
-
-        // Determine which name to search for in the target file, based on
-        // what this import makes visible in the current scope.
-        let search_name: &str = match &import.spec {
-            Some(ImportSpec::Names(names)) => {
-                // Find the entry whose visible name (alias if present, otherwise
-                // the original name) matches `word`.
-                let Some(entry) = names
-                    .iter()
-                    .find(|n| n.alias.as_deref().unwrap_or(n.name.as_str()) == word)
-                else {
-                    continue; // this import does not bring `word` into scope
-                };
-                &entry.name // search target file by the *original* name
-            }
-            Some(ImportSpec::Glob) => word, // glob: any name may come from here
-            None => {
-                // Bare path import (`import foo::bar`).  The last path segment
-                // is itself a navigable target (e.g. cursor on `bar`).
-                if import.path.last().map(String::as_str) == Some(word)
-                    && (documents.contains_key(&target_uri) || path.exists())
-                {
-                    return Some((target_uri, Range::default()));
-                }
-                continue;
-            }
-        };
-
-        // Search in an already-open document first (no I/O).
-        if let Some(doc) = documents.get(&target_uri) {
-            if let Some(range) = find_definition_in_ast(
-                &doc.source,
-                &doc.line_offsets,
-                &doc.parse_result,
-                search_name,
-            ) {
-                return Some((target_uri, range));
-            }
-            // File is open but name not found there — don't fall through to disk.
-            continue;
-        }
-
-        // Fall back: read and parse the file from disk.
-        if path.exists() {
-            if let Ok(source) = std::fs::read_to_string(&path) {
-                let pr = hew_parser::parse(&source);
-                let lo = compute_line_offsets(&source);
-                if let Some(range) = find_definition_in_ast(&source, &lo, &pr, search_name) {
-                    return Some((target_uri, range));
-                }
-            }
-        }
-    }
-    None
-}
-
-// ── Document links ──────────────────────────────────────────────────
-
-fn build_document_links(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-) -> Vec<DocumentLink> {
-    let mut links = Vec::new();
-
-    for (item, span) in &parse_result.program.items {
-        if let Item::Import(import) = item {
-            let Some(path) = compute_import_path(uri, import) else {
-                continue;
-            };
-            if !path.exists() {
-                continue;
-            }
-            if let Ok(target_uri) = Url::from_file_path(&path) {
-                let relative = format!("{}.hew", import.path.join("/"));
-                links.push(DocumentLink {
-                    range: span_to_range(source, lo, span),
-                    target: Some(target_uri),
-                    tooltip: Some(format!("Open {relative}")),
-                    data: None,
-                });
-            }
-        }
-    }
-    links
-}
-
-// ── Type hierarchy helpers ──────────────────────────────────────────
-
-fn find_type_hierarchy_item(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    name: &str,
-) -> Option<TypeHierarchyItem> {
-    for (item, span) in &parse_result.program.items {
-        let (item_name, kind) = match item {
-            Item::TypeDecl(td) => {
-                let kind = match td.kind {
-                    TypeDeclKind::Struct => SymbolKind::STRUCT,
-                    TypeDeclKind::Enum => SymbolKind::ENUM,
-                };
-                (td.name.as_str(), kind)
-            }
-            Item::Actor(a) => (a.name.as_str(), SymbolKind::CLASS),
-            Item::Trait(t) => (t.name.as_str(), SymbolKind::INTERFACE),
-            _ => continue,
-        };
-        if item_name == name {
-            let range = span_to_range(source, lo, span);
-            return Some(TypeHierarchyItem {
-                name: item_name.to_string(),
-                kind,
-                tags: None,
-                detail: None,
-                uri: uri.clone(),
-                range,
-                selection_range: range,
-                data: None,
-            });
-        }
-    }
-    None
-}
-
-fn collect_supertypes(
-    uri: &Url,
-    name: &str,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-) -> Vec<TypeHierarchyItem> {
-    let mut supers = Vec::new();
-
-    for (item, _) in &parse_result.program.items {
-        match item {
-            // For actors, collect their declared super_traits and return immediately.
-            Item::Actor(a) if a.name == name => {
-                if let Some(bounds) = &a.super_traits {
-                    for bound in bounds {
-                        if let Some(hi) =
-                            find_type_hierarchy_item(uri, source, lo, parse_result, &bound.name)
-                        {
-                            supers.push(hi);
-                        }
-                    }
-                }
-                return supers;
-            }
-            // For types, find impl blocks: `impl TraitName for TypeName`
-            Item::Impl(impl_decl) => {
-                let target = match &impl_decl.target_type.0 {
-                    hew_parser::ast::TypeExpr::Named { name: tname, .. } => tname.as_str(),
-                    _ => continue,
-                };
-                if target == name {
-                    if let Some(bound) = &impl_decl.trait_bound {
-                        if let Some(hi) =
-                            find_type_hierarchy_item(uri, source, lo, parse_result, &bound.name)
-                        {
-                            supers.push(hi);
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    supers
-}
-
-fn collect_subtypes(
-    uri: &Url,
-    trait_name: &str,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-) -> Vec<TypeHierarchyItem> {
-    let mut subs = Vec::new();
-
-    // Find types that implement this trait via `impl TraitName for TypeName`.
-    for (item, _) in &parse_result.program.items {
-        if let Item::Impl(impl_decl) = item {
-            if let Some(bound) = &impl_decl.trait_bound {
-                if bound.name == trait_name {
-                    let target = match &impl_decl.target_type.0 {
-                        hew_parser::ast::TypeExpr::Named { name: tname, .. } => tname.as_str(),
-                        _ => continue,
-                    };
-                    if let Some(hi) =
-                        find_type_hierarchy_item(uri, source, lo, parse_result, target)
-                    {
-                        subs.push(hi);
-                    }
-                }
-            }
-        }
-    }
-
-    // Find actors that declare this trait as a super_trait.
-    for (item, _) in &parse_result.program.items {
-        if let Item::Actor(a) = item {
-            if let Some(bounds) = &a.super_traits {
-                if bounds.iter().any(|b| b.name == trait_name) {
-                    if let Some(hi) =
-                        find_type_hierarchy_item(uri, source, lo, parse_result, &a.name)
-                    {
-                        subs.push(hi);
-                    }
-                }
-            }
-        }
-    }
-    subs
-}
-
-// ── Call hierarchy helpers ──────────────────────────────────────────
-
-fn find_callable_at(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    name: &str,
-) -> Option<CallHierarchyItem> {
-    for (item, item_span) in &parse_result.program.items {
-        match item {
-            Item::Function(f) if f.name == name => {
-                let range = span_to_range(source, lo, item_span);
-                return Some(CallHierarchyItem {
-                    name: f.name.clone(),
-                    kind: SymbolKind::FUNCTION,
-                    tags: None,
-                    detail: None,
-                    uri: uri.clone(),
-                    range,
-                    selection_range: range,
-                    data: None,
-                });
-            }
-            Item::Actor(a) => {
-                for recv in &a.receive_fns {
-                    if recv.name == name {
-                        let fn_span = if recv.span.is_empty() {
-                            item_span
-                        } else {
-                            &recv.span
-                        };
-                        let range = span_to_range(source, lo, fn_span);
-                        return Some(CallHierarchyItem {
-                            name: recv.name.clone(),
-                            kind: SymbolKind::METHOD,
-                            tags: None,
-                            detail: Some(format!("actor {}", a.name)),
-                            uri: uri.clone(),
-                            range,
-                            selection_range: range,
-                            data: None,
-                        });
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    None
-}
-
-#[expect(
-    clippy::too_many_lines,
-    reason = "exhaustive match over caller item kinds is clearest as one function"
-)]
-fn find_incoming_calls(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    target_name: &str,
-) -> Vec<CallHierarchyIncomingCall> {
-    let mut result = Vec::new();
-    for (item, item_span) in &parse_result.program.items {
-        // Collect all call sites inside this item using the exhaustive item walker.
-        let mut calls = Vec::new();
-        collect_calls_in_item(item, &mut calls);
-        let matching: Vec<_> = calls.iter().filter(|c| c.name == target_name).collect();
-        if matching.is_empty() {
-            continue;
-        }
-        let from_ranges: Vec<_> = matching
-            .iter()
-            .map(|c| span_to_range(source, lo, &c.span))
-            .collect();
-
-        // Derive display metadata from the item kind.
-        match item {
-            Item::Function(f) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: f.name.clone(),
-                        kind: SymbolKind::FUNCTION,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::Actor(a) => {
-                // Walk each named body independently for precise attribution.
-                // init body (no dedicated name span — fall back to item span).
-                if let Some(init) = &a.init {
-                    let mut body_calls = Vec::new();
-                    collect_calls_in_block(&init.body, &mut body_calls);
-                    let fn_calls: Vec<_> = body_calls
-                        .iter()
-                        .filter(|c| c.name == target_name)
-                        .collect();
-                    if !fn_calls.is_empty() {
-                        let range = span_to_range(source, lo, item_span);
-                        result.push(CallHierarchyIncomingCall {
-                            from: CallHierarchyItem {
-                                name: format!("{}.init", a.name),
-                                kind: SymbolKind::METHOD,
-                                tags: None,
-                                detail: None,
-                                uri: uri.clone(),
-                                range,
-                                selection_range: range,
-                                data: None,
-                            },
-                            from_ranges: fn_calls
-                                .iter()
-                                .map(|c| span_to_range(source, lo, &c.span))
-                                .collect(),
-                        });
-                    }
-                }
-                // terminate body.
-                if let Some(term) = &a.terminate {
-                    let mut body_calls = Vec::new();
-                    collect_calls_in_block(&term.body, &mut body_calls);
-                    let fn_calls: Vec<_> = body_calls
-                        .iter()
-                        .filter(|c| c.name == target_name)
-                        .collect();
-                    if !fn_calls.is_empty() {
-                        let range = span_to_range(source, lo, item_span);
-                        result.push(CallHierarchyIncomingCall {
-                            from: CallHierarchyItem {
-                                name: format!("{}.terminate", a.name),
-                                kind: SymbolKind::METHOD,
-                                tags: None,
-                                detail: None,
-                                uri: uri.clone(),
-                                range,
-                                selection_range: range,
-                                data: None,
-                            },
-                            from_ranges: fn_calls
-                                .iter()
-                                .map(|c| span_to_range(source, lo, &c.span))
-                                .collect(),
-                        });
-                    }
-                }
-                // Receive functions.
-                for recv in &a.receive_fns {
-                    let mut body_calls = Vec::new();
-                    collect_calls_in_block(&recv.body, &mut body_calls);
-                    let fn_calls: Vec<_> = body_calls
-                        .iter()
-                        .filter(|c| c.name == target_name)
-                        .collect();
-                    if fn_calls.is_empty() {
-                        continue;
-                    }
-                    let fn_span = if recv.span.is_empty() {
-                        item_span
-                    } else {
-                        &recv.span
-                    };
-                    let range = span_to_range(source, lo, fn_span);
-                    result.push(CallHierarchyIncomingCall {
-                        from: CallHierarchyItem {
-                            name: recv.name.clone(),
-                            kind: SymbolKind::METHOD,
-                            tags: None,
-                            detail: Some(format!("actor {}", a.name)),
-                            uri: uri.clone(),
-                            range,
-                            selection_range: range,
-                            data: None,
-                        },
-                        from_ranges: fn_calls
-                            .iter()
-                            .map(|c| span_to_range(source, lo, &c.span))
-                            .collect(),
-                    });
-                }
-                // Actor methods.
-                for method in &a.methods {
-                    let mut body_calls = Vec::new();
-                    collect_calls_in_block(&method.body, &mut body_calls);
-                    let fn_calls: Vec<_> = body_calls
-                        .iter()
-                        .filter(|c| c.name == target_name)
-                        .collect();
-                    if fn_calls.is_empty() {
-                        continue;
-                    }
-                    let range = if method.decl_span.is_empty() {
-                        span_to_range(source, lo, item_span)
-                    } else {
-                        span_to_range(source, lo, &method.decl_span)
-                    };
-                    result.push(CallHierarchyIncomingCall {
-                        from: CallHierarchyItem {
-                            name: method.name.clone(),
-                            kind: SymbolKind::METHOD,
-                            tags: None,
-                            detail: Some(format!("actor {}", a.name)),
-                            uri: uri.clone(),
-                            range,
-                            selection_range: range,
-                            data: None,
-                        },
-                        from_ranges: fn_calls
-                            .iter()
-                            .map(|c| span_to_range(source, lo, &c.span))
-                            .collect(),
-                    });
-                }
-            }
-            Item::Impl(i) => {
-                let impl_name = match &i.target_type.0 {
-                    hew_parser::ast::TypeExpr::Named { name: tname, .. } => tname.clone(),
-                    _ => "<impl>".to_string(),
-                };
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: impl_name,
-                        kind: SymbolKind::NAMESPACE,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::TypeDecl(td) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: td.name.clone(),
-                        kind: SymbolKind::STRUCT,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::Trait(t) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: t.name.clone(),
-                        kind: SymbolKind::INTERFACE,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::Supervisor(s) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: s.name.clone(),
-                        kind: SymbolKind::MODULE,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::Machine(m) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: m.name.clone(),
-                        kind: SymbolKind::ENUM,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            Item::Const(c) => {
-                let range = span_to_range(source, lo, item_span);
-                result.push(CallHierarchyIncomingCall {
-                    from: CallHierarchyItem {
-                        name: c.name.clone(),
-                        kind: SymbolKind::CONSTANT,
-                        tags: None,
-                        detail: None,
-                        uri: uri.clone(),
-                        range,
-                        selection_range: range,
-                        data: None,
-                    },
-                    from_ranges,
-                });
-            }
-            _ => {}
-        }
-    }
-    result
-}
-
-fn find_outgoing_calls(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    caller_name: &str,
-) -> Vec<CallHierarchyOutgoingCall> {
-    let mut call_sites = Vec::new();
-
-    // Collect calls only from the specific body that matches caller_name.
-    // For multi-body items (Actor, Impl, TypeDecl, Trait) this walks only the
-    // matching sub-body, preventing sibling methods from bleeding into the
-    // outgoing call set.
-    for (item, _) in &parse_result.program.items {
-        collect_calls_in_named_body(item, caller_name, &mut call_sites);
-    }
-
-    // Group call sites by callee name
-    let mut grouped: HashMap<String, Vec<Range>> = HashMap::new();
-    for cs in &call_sites {
-        grouped
-            .entry(cs.name.clone())
-            .or_default()
-            .push(span_to_range(source, lo, &cs.span));
-    }
-
-    grouped
-        .into_iter()
-        .filter_map(|(callee_name, ranges)| {
-            let target = find_callable_at(uri, source, lo, parse_result, &callee_name)?;
-            Some(CallHierarchyOutgoingCall {
-                to: target,
-                from_ranges: ranges,
-            })
-        })
-        .collect()
-}
-
-// ── Code lens helpers ───────────────────────────────────────────────
-
-fn build_code_lenses(source: &str, lo: &[usize], parse_result: &ParseResult) -> Vec<CodeLens> {
-    let ref_counts = count_all_references(parse_result);
-    let mut lenses = Vec::new();
-
-    let ref_lens = |range, name: &str| -> CodeLens {
-        let count = ref_counts.get(name).copied().unwrap_or(0);
-        CodeLens {
-            range,
-            command: Some(Command {
-                title: format!("{count} reference{}", if count == 1 { "" } else { "s" }),
-                command: String::new(),
-                arguments: None,
-            }),
-            data: None,
-        }
-    };
-
-    for (item, item_span) in &parse_result.program.items {
-        match item {
-            Item::Function(f) => {
-                let range = span_to_range(source, lo, item_span);
-                lenses.push(ref_lens(range, &f.name));
-                if has_test_attribute(&f.attributes) {
-                    lenses.push(CodeLens {
-                        range,
-                        command: Some(Command {
-                            title: "\u{25b6} Run test".to_string(),
-                            command: "hew.runTest".to_string(),
-                            arguments: Some(vec![serde_json::Value::String(f.name.clone())]),
-                        }),
-                        data: None,
-                    });
-                }
-            }
-            Item::Actor(a) => {
-                let range = span_to_range(source, lo, item_span);
-                lenses.push(ref_lens(range, &a.name));
-                for recv in &a.receive_fns {
-                    let recv_range = if recv.span.is_empty() {
-                        range
-                    } else {
-                        span_to_range(source, lo, &recv.span)
-                    };
-                    lenses.push(ref_lens(recv_range, &recv.name));
-                }
-            }
-            _ => {}
-        }
-    }
-    lenses
-}
-
-fn has_test_attribute(attrs: &[Attribute]) -> bool {
-    attrs.iter().any(|a| a.name == "test")
-}
-
-// ── Workspace symbol helpers ────────────────────────────────────────
-
-#[expect(
-    deprecated,
-    reason = "SymbolInformation::deprecated field is deprecated in lsp-types"
-)]
-#[expect(
-    clippy::too_many_lines,
-    reason = "exhaustive match over all Item variants for workspace symbols"
-)]
-fn collect_workspace_symbols(
-    uri: &Url,
-    source: &str,
-    lo: &[usize],
-    parse_result: &ParseResult,
-    query: &str,
-) -> Vec<SymbolInformation> {
-    let mut symbols = Vec::new();
-    let query_lower = query.to_lowercase();
-    for (item, item_span) in &parse_result.program.items {
-        match item {
-            Item::Function(f) => {
-                if query.is_empty() || f.name.to_lowercase().contains(&query_lower) {
-                    symbols.push(SymbolInformation {
-                        name: f.name.clone(),
-                        kind: SymbolKind::FUNCTION,
-                        tags: None,
-                        deprecated: None,
-                        location: Location {
-                            uri: uri.clone(),
-                            range: span_to_range(source, lo, item_span),
-                        },
-                        container_name: None,
-                    });
-                }
-            }
-            Item::Actor(a) => {
-                if query.is_empty() || a.name.to_lowercase().contains(&query_lower) {
-                    symbols.push(SymbolInformation {
-                        name: a.name.clone(),
-                        kind: SymbolKind::CLASS,
-                        tags: None,
-                        deprecated: None,
-                        location: Location {
-                            uri: uri.clone(),
-                            range: span_to_range(source, lo, item_span),
-                        },
-                        container_name: None,
-                    });
-                }
-                for recv in &a.receive_fns {
-                    if query.is_empty() || recv.name.to_lowercase().contains(&query_lower) {
-                        let recv_range = if recv.span.is_empty() {
-                            span_to_range(source, lo, item_span)
-                        } else {
-                            span_to_range(source, lo, &recv.span)
-                        };
-                        symbols.push(SymbolInformation {
-                            name: recv.name.clone(),
-                            kind: SymbolKind::METHOD,
-                            tags: None,
-                            deprecated: None,
-                            location: Location {
-                                uri: uri.clone(),
-                                range: recv_range,
-                            },
-                            container_name: Some(a.name.clone()),
-                        });
-                    }
-                }
-            }
-            Item::TypeDecl(t) => {
-                if query.is_empty() || t.name.to_lowercase().contains(&query_lower) {
-                    symbols.push(SymbolInformation {
-                        name: t.name.clone(),
-                        kind: SymbolKind::STRUCT,
-                        tags: None,
-                        deprecated: None,
-                        location: Location {
-                            uri: uri.clone(),
-                            range: span_to_range(source, lo, item_span),
-                        },
-                        container_name: None,
-                    });
-                }
-            }
-            Item::Const(c) => {
-                if query.is_empty() || c.name.to_lowercase().contains(&query_lower) {
-                    symbols.push(SymbolInformation {
-                        name: c.name.clone(),
-                        kind: SymbolKind::CONSTANT,
-                        tags: None,
-                        deprecated: None,
-                        location: Location {
-                            uri: uri.clone(),
-                            range: span_to_range(source, lo, item_span),
-                        },
-                        container_name: None,
-                    });
-                }
-            }
-            Item::Trait(t) => {
-                if query.is_empty() || t.name.to_lowercase().contains(&query_lower) {
-                    symbols.push(SymbolInformation {
-                        name: t.name.clone(),
-                        kind: SymbolKind::INTERFACE,
-                        tags: None,
-                        deprecated: None,
-                        location: Location {
-                            uri: uri.clone(),
-                            range: span_to_range(source, lo, item_span),
-                        },
-                        container_name: None,
-                    });
-                }
-            }
-            _ => {}
-        }
-    }
-    symbols
 }
 
 #[cfg(test)]
